Time: 0 ms (100.00%) | Memory: 8.3 MB (12.22%) - LeetSync

+++ 7-reverse-integer/reverse-integer.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    int reverse(int x) {
+    long long reverse(int x) {
         long long rev = 0;
     while(x)
         {
